backend/Django/requirements.txt
Django==2.0
psycopg2-binary
djoser
djangorestframework==3.11.0
social-auth-app-django
djangorestframework-simplejwt
rest-social-auth
django-cors-headers
django-guardian
pydicom
pyotp
cryptography
docker
kubernetes
gevent
gunicorn
dicomweb-client
scikit-build
monailabel